renaming and cleanup and arch diagram
