+++ worker/README.md
@@ -1,0 +1,3 @@
+# Below is a high level overview describing how the worker is architected
+
+<img src="./assets/worker-arch.png">
